Fix int casting following PR comment

+++ host-linux/Signer.h
@@ -210,7 +210,7 @@ private:
             pin->setValidator(new QRegExpValidator(QRegExp(QString("\\d{%1,12}").arg(minPinLen)), pin));
             pin->setMaxLength(12);
             connect(pin, &QLineEdit::textEdited, [=](const QString &text){
-                ok->setEnabled((unsigned long)text.size() >= minPinLen);
+                ok->setEnabled(text.size() >= int(minPinLen));
             });
 
             layout->addWidget(pin);
